+++ src/main/java/abc/company/plt/employee/managment/system/service/EmployeeService.java
@@ -16,6 +16,9 @@ public class EmployeeService {
 
     // Find one employee by ID
     public Optional<Employee> getEmployeeById(Long employeeID){
+        if(employeeID==null || employeeID<=0){
+            throw new IllegalArgumentException("Invalid Employee ID");
+        }
         return employeeRepository.findById(employeeID);
     }
 
@@ -26,26 +29,34 @@ public class EmployeeService {
 
     // Delete an employee by ID
     public void deleteEmployeeById(Long id){
+        if(id==null || id<=0){
+            throw new IllegalArgumentException("Invalid Employee ID");
+        }
         employeeRepository.deleteById(id);
     }
 
     // Add a new employee
     public Employee addEmployee(Employee employee){
+        if (employee == null || employee.getEmail() == null || employee.getEmail().isEmpty()) {
+            throw new IllegalArgumentException("Employee email cannot be null or empty");
+        }
         return employeeRepository.save(employee);
     }
 
     // Update an existing employee
     public Employee updateEmployee(Long employeeId, Employee employeeDetails){
+        if(employeeId==null || employeeId<=0){
+            throw new IllegalArgumentException("Invalid Employee ID");
+        }
         Employee employee = employeeRepository.findById(employeeId)
                 .orElseThrow(() -> new RuntimeException("Employee not found with id: " + employeeId));
 
         // Update fields from the details provided
         employee.setEmail(employeeDetails.getEmail());
-        employee.setPassword(employeeDetails.getPassword()); // Consider password handling
+        employee.setPassword(employeeDetails.getPassword()); 
         employee.setFirstName(employeeDetails.getFirstName());
         employee.setLastName(employeeDetails.getLastName());
-        employee.setDepartment(employeeDetails.getDepartment()); // Update associated department
-
+        employee.setDepartment(employeeDetails.getDepartment());
         return employeeRepository.save(employee);
     }
